refactor(voice/capture): extract exception classes into capture subpackage [T1.4 step 2]
Master mission Phase 1 / T1.4 — second stage of the
``_capture_task.py`` split. Builds on step 1 (`32efb40`,
restart-verdict types) by extracting the structured exception
hierarchy. Same low-risk profile: pure types, no class-state
coupling, no behaviour change.

What lands:

  * New module ``capture/_exceptions.py`` containing the four
    capture exception classes:
      - ``CaptureError`` — base (RuntimeError subclass; isinstance
        check-point for the dashboard ``/api/voice/enable`` route).
      - ``CaptureSilenceError`` — opened, all-zero frames.
      - ``CaptureInoperativeError`` — cascade exhausted.
      - ``CaptureDeviceContendedError`` — Linux session-manager
        contention.
    Total ~135 LOC. ``OpenAttempt`` typed-only import preserved
    via ``TYPE_CHECKING`` (matches the source pattern; runtime
    cost zero).
  * ``_capture_task.py`` drops the four class definitions, adds
    ``import X as X`` re-exports for each name (mypy strict
    requires this form for an import to count as a public
    re-export).
  * ``capture/__init__.py`` re-exports the four exception
    names alongside the step-1 restart types.

External call sites that import these exceptions (dashboard
``test_voice_routes.py``, ``test_voice_factory.py``,
integration ``test_voice_recovery_hybrid.py``) keep working
unchanged via the re-export at the legacy path.

T1.4 step-1 + step-2 cumulative impact:

  * ``_capture_task.py``: 2785 → ~2245 LOC (~540 LOC extracted,
    ~19 % of the split done).
  * Two coherent sub-modules under ``voice/capture/`` with
    pure-types content. The remaining ~2245 LOC — ``AudioCaptureTask``
    class itself + module-level constants + low-level helpers —
    is the higher-risk surgery the spec calls "DEDICATED SESSION".
    Each future step can land independently with the legacy stub
    intact.

Validation:
  * ruff check + format clean
  * mypy strict: 0 issues across 421 source files (+1 source file
    for ``capture/_exceptions.py``)
  * bandit clean
  * pytest full suite: 11894 passed, 25 skipped, 57 deselected
    (zero delta — pure structural extraction, no test changes
    needed because legacy import surface is preserved)

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 (step 2)
